+++ templates/slides/app/components/editor/EditorToolbar.tsx
@@ -31,7 +31,7 @@ import {
   IconPin,
   IconLetterT,
   IconWand,
-  IconUpload,
+  IconDownload,
   IconSun,
   IconMoon,
   IconDotsVertical,
@@ -1084,7 +1084,7 @@ graph TD
             {importing ? (
               <IconLoader2 className="w-4 h-4 mr-2 animate-spin" />
             ) : (
-              <IconUpload className="w-4 h-4 mr-2" />
+              <IconDownload className="w-4 h-4 mr-2" />
             )}
             {importing
               ? t("editorToolbar.importing")